+++ src/AppForSEII2526.API/DTOs/ReseñasDTOs/ReviewDTO.cs
@@ -3,16 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-    public class ReviewItemDetailsDTO
-    {
-        public string DeviceName { get; set; }
-        public string DeviceModel { get; set; }
-        public int DeviceYear { get; set; }
-        public int Rating { get; set; } 
-        public string? Comment { get; set; } 
-    }
-
-    
     public class ReviewDetailsDTO
     {
         public string Username { get; set; }
